# Skip header validation for data parsed off the wire