The parser already validates, so there's no need to do it twice, and
it's a bit expensive.

On my laptop, on the benchmark in bench/, I get:
  CPython 3.5: ~4386 req/sec -> ~5141 req/sec = 17% speedup
  PyPy 5.6.0: ~19311 req/sec -> ~21106 req/sec = 9% speedup